requirements.txt
colorama==0.4.6
evdev==1.6.1
Pillow==9.4.0
pynput==1.7.6
PyYAML==6.0
